refactor: SIMT-16 code refactoring
