+++ feature/feature-home/src/main/java/com/comit/feature_home/screen/HomeScreen.kt
@@ -67,7 +67,7 @@ fun HomeScreen() {
 
                 Image(
                     painter = painterResource(id = R.drawable.ic_home_next),
-                    contentDescription = "",
+                    contentDescription = null,
                     modifier = Modifier
                         .fillMaxHeight()
                         .wrapContentHeight(Alignment.CenterVertically)
